+++ Soulgram.Interests/Soulgram.Interests.Api/Controllers/GenresController.cs
@@ -5,6 +5,7 @@ using Soulgram.Interests.Application.Commands.Genres;
 using Soulgram.Interests.Application.Models.Request;
 using Soulgram.Interests.Application.Models.Response;
 using Soulgram.Interests.Application.Queries;
+using Soulgram.Interests.Application.Queries.Genres;
 
 namespace Soulgram.Interests.Api.Controllers;
