+++ Soulgram.Interests/Soulgram.Interests.Application/Queries/Interests/GetInterestsForUserQuery.cs
@@ -2,7 +2,7 @@ using MediatR;
 using Soulgram.Interests.Application.Interfaces;
 using Soulgram.Interests.Application.Models.Response;
 
-namespace Soulgram.Interests.Application.Queries;
+namespace Soulgram.Interests.Application.Queries.Interests;
 
 public class GetInterestsForUserQuery : IRequest<IEnumerable<InterestResponse>>
 {
