+++ Soulgram.Interests/Soulgram.Interests.Application/Queries/Movies/GetUserMoviesQueryHandler.cs
@@ -3,7 +3,7 @@ using Soulgram.Interests.Application.Interfaces;
 using Soulgram.Interests.Application.Models.Response;
 using Soulgram.Interests.Domain;
 
-namespace Soulgram.Interests.Application.Queries;
+namespace Soulgram.Interests.Application.Queries.Movies;
 
 public class GetUserMoviesQueryHandler : IRequestHandler<GetUserMoviesQuery, ICollection<MovieSearchResponse>?>
 {
