+++ Soulgram.Interests/Soulgram.Interests.Infrastructure/Facades/IMovieFacade.cs
@@ -6,6 +6,6 @@ namespace Soulgram.Interests.Infrastructure.Facades;
 public interface IMovieFacade
 {
     Task<ICollection<string>?> GetGenresAsync(CancellationToken cancellationToken);
-    Task<IEnumerable<MovieSearchResponse>> GetMoviesByName(SearchMovieRequest request,
+    Task<IEnumerable<MovieSearchResponse>?> GetMoviesByName(SearchMovieRequest request,
         CancellationToken cancellationToken);
 }
