+++ Soulgram.Interests/Soulgram.Interests.Infrastructure/Filters/IMovieResponseFilter.cs
@@ -5,5 +5,5 @@ namespace Soulgram.Interests.Infrastructure.Filters;
 
 public interface IMovieResponseFilter
 {
-    public IEnumerable<MovieSearchResponse> Filter(int limit, IEnumerable<MovieSearchResponse> models);
+    public IEnumerable<MovieSearchResponse> Filter(int limit, IEnumerable<MovieSearchResponse>? models);
 }
